alerts/requirements.txt
PyYAML==5.3.1
urllib3==1.25.8
requests==2.26.0
